♻  refactor: Move `module_test_path` fixture to `conftest`
